fix(twitter): respect routes to non-user pages
Prior to this commit, the Twitter user resolver attempted to resolve
pages such as
    - twitter.com/home or
    - twitter.com/explore
as a user page. This lead to 404s and 403s. With this commit, we define
a collection of known non-user pages that will not be considered as user
pages when encountered.

The list of pages was sourced by looking around the Twitter web
interface so it is probably not exhaustive. If you find additional
special pages, please append them to `nonUserPages`.